Move files into Copilot CLI specific folders
