+++ extensions/copilot/src/extension/chatSessions/copilotcli/vscode-node/copilotCLIChatSessionInitializer.ts
@@ -6,21 +6,21 @@
 import type { SweCustomAgent } from '@github/copilot/sdk';
 import * as l10n from '@vscode/l10n';
 import * as vscode from 'vscode';
-import { ConfigKey, IConfigurationService } from '../../../platform/configuration/common/configurationService';
-import { ILogService } from '../../../platform/log/common/logService';
-import { IPromptsService, ParsedPromptFile } from '../../../platform/promptFiles/common/promptsService';
-import { IWorkspaceService } from '../../../platform/workspace/common/workspaceService';
-import { createServiceIdentifier } from '../../../util/common/services';
-import { DisposableStore, IReference } from '../../../util/vs/base/common/lifecycle';
-import { URI } from '../../../util/vs/base/common/uri';
-import { ChatVariablesCollection, extractDebugTargetSessionIds, isPromptFile } from '../../prompt/common/chatVariablesCollection';
-import { FolderRepositoryInfo, IFolderRepositoryManager, IsolationMode } from '../common/folderRepositoryManager';
-import { emptyWorkspaceInfo, getWorkingDirectory, isIsolationEnabled, IWorkspaceInfo } from '../common/workspaceInfo';
-import { SessionIdForCLI } from '../copilotcli/common/utils';
-import { COPILOT_CLI_REASONING_EFFORT_PROPERTY, ICopilotCLIAgents, ICopilotCLIModels } from '../copilotcli/node/copilotCli';
-import { ICopilotCLISession } from '../copilotcli/node/copilotcliSession';
-import { ICopilotCLISessionService } from '../copilotcli/node/copilotcliSessionService';
-import { buildMcpServerMappings, McpServerMappings } from '../copilotcli/node/mcpHandler';
+import { ConfigKey, IConfigurationService } from '../../../../platform/configuration/common/configurationService';
+import { ILogService } from '../../../../platform/log/common/logService';
+import { IPromptsService, ParsedPromptFile } from '../../../../platform/promptFiles/common/promptsService';
+import { IWorkspaceService } from '../../../../platform/workspace/common/workspaceService';
+import { createServiceIdentifier } from '../../../../util/common/services';
+import { DisposableStore, IReference } from '../../../../util/vs/base/common/lifecycle';
+import { URI } from '../../../../util/vs/base/common/uri';
+import { ChatVariablesCollection, extractDebugTargetSessionIds, isPromptFile } from '../../../prompt/common/chatVariablesCollection';
+import { FolderRepositoryInfo, IFolderRepositoryManager, IsolationMode } from '../../common/folderRepositoryManager';
+import { emptyWorkspaceInfo, getWorkingDirectory, isIsolationEnabled, IWorkspaceInfo } from '../../common/workspaceInfo';
+import { SessionIdForCLI } from '../../copilotcli/common/utils';
+import { COPILOT_CLI_REASONING_EFFORT_PROPERTY, ICopilotCLIAgents, ICopilotCLIModels } from '../../copilotcli/node/copilotCli';
+import { ICopilotCLISession } from '../../copilotcli/node/copilotcliSession';
+import { ICopilotCLISessionService } from '../../copilotcli/node/copilotcliSessionService';
+import { buildMcpServerMappings, McpServerMappings } from '../../copilotcli/node/mcpHandler';
 
 function isReasoningEffortFeatureEnabled(configurationService: IConfigurationService): boolean {
 	return configurationService.getConfig(ConfigKey.Advanced.CLIThinkingEffortEnabled);
