+++ extensions/copilot/src/extension/chatSessions/copilotcli/vscode-node/test/copilotCLISDKUpgrade.spec.ts
@@ -7,8 +7,8 @@ import { promises as fs } from 'fs';
 import { isBinaryFile } from 'isbinaryfile';
 import * as path from 'path';
 import { beforeAll, describe, it } from 'vitest';
-import { TestLogService } from '../../../../platform/testing/common/testLogService';
-import { copyRipgrepShim } from '../../copilotcli/node/ripgrepShim';
+import { TestLogService } from '../../../../../platform/testing/common/testLogService';
+import { copyRipgrepShim } from '../../../copilotcli/node/ripgrepShim';
 
 describe('CopilotCLI SDK Upgrade', function () {
 	const extensionPath = path.join(__dirname, '..', '..', '..', '..', '..');
